pull addCommand from aliases spec into utils
